requirements.txt
scrapy
scrapy-user-agents
Pillow
psycopg2-binary